requirements.txt
MouseInfo==0.1.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
PyAutoGUI==0.9.35
PyGetWindow==0.0.9
PyMsgBox==1.0.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.8
PyTweening==1.0.3
pywin32==301
pyzmq==19.0.0
zmq==0.0.0
